change updater

+++ ui/logfile.py
@@ -7,14 +7,18 @@ logobj = None
 
 class LogFile:
     def __init__(self, filename, charset='utf-8'):
-        self.filename = filename
         self.charset = charset
-        if os.path.isfile(filename):
-            bakfile = filename + '.bak'
-            if os.path.isfile(bakfile):
-                os.remove(bakfile)
-            shutil.move(filename, bakfile)
-        self.file = open(filename, 'a+')
+        self.filename = filename
+        
+        if filename == 'stdout':
+            self.file = sys.stdout
+        else:
+            if os.path.isfile(filename):
+                bakfile = filename + '.bak'
+                if os.path.isfile(bakfile):
+                    os.remove(bakfile)
+                shutil.move(filename, bakfile)
+            self.file = open(filename, 'a+')
 
     def info(self, *s):
         ss = []
